% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaner.R
\name{absolute.calculator}
\alias{absolute.calculator}
\title{absolute.calculator()}
\usage{
absolute.calculator(data, absolute.dataset, qc = 1:5)
}
\arguments{
\item{data}{Row data of lipidomics}

\item{absolute.dataset}{Data frame. Normalization index.}

\item{qc}{Numeric vector. The column of quality control.}
}
\value{
Return a data frame of absolute normalized lipidomic data.
}
\description{
Internal function of lipidmicR::normalization_calculator()
}
