% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance.signif}
\alias{abundance.signif}
\title{Analyze Abundance Significance}
\usage{
abundance.signif(
  data,
  group,
  istotal = FALSE,
  by = data.frame(group = c(), variable.1 = c(), variable.2 = c())
)
}
\arguments{
\item{data}{The data frame containing abundance data.}

\item{group}{A vector specifying the group membership for each sample.}

\item{istotal}{Logical. If is true, statistics based on total summary table of lipid type, carbon number and unsaturation rate will be generated.}

\item{by}{A data frame specifying additional variables for the analysis.}
}
\value{
A list containing statistical analysis results for each species.
}
\description{
This function performs statistical analysis to determine the significance of abundance data.
}
