% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaner.R
\name{cleanXpert}
\alias{cleanXpert}
\title{cleanXpert()}
\usage{
cleanXpert(interactive = FALSE, parameters = NULL)
}
\arguments{
\item{interactive}{Logical input indicating whether to use the paramWidzard interface.}

\item{parameters}{List input including the parameters for data cleaning. Only useful when interactive == FALSE.}
}
\value{
A list containing cleaned data with or without annotation.
}
\description{
Integrated function for lipidomic data cleaning, processing and normalization.
}
