% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importer_group.R
\name{groupXpert}
\alias{groupXpert}
\title{groupXpert()}
\usage{
groupXpert(data, sep = "_", as.name = TRUE, specify = NULL)
}
\arguments{
\item{data}{Data.frame. Row data to be grouped.}

\item{sep}{Character. The separator used to separate variables from the number of repeats}

\item{as.name}{Logical. If is true, the group name will be used as the name of the return value. If is false, the sample name will be used as the name of the return value.}

\item{specify}{List. Used to set groups manually. Each sublist has a group name and a value based on the column range in which it is located. No duplicate values can appear in the sublist.}
}
\value{
A vector named as the colname and uses group name as the value.
}
\description{
groupXpert()
}
