% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaner.R
\name{lEa}
\alias{lEa}
\title{lEa()}
\usage{
lEa(
  data,
  Inchannel = c("PC(15:0/18:1(d7))+AcO_1", "PE(15:0/18:1(d7))_2", "PS(15:0/18:1(d7))_2",
    "PG(15:0/18:1(d7))_2", "PI(15:0/18:1(d7))_2", "PA(15:0/18:1(d7))_2",
    "LPC18:1(d7)+AcO", "LPE18:1(d7)"),
  sample = 2:5
)
}
\arguments{
\item{data}{Data frame. Row lipidomics data. Should be imported with importer().}

\item{Inchannel}{Vector. Exact channel name for isotope label data.}

\item{sample}{Vector. Column of sample added with isotope label. Default as 2:5.}
}
\value{
Data of internal label.
}
\description{
Internal function to extract label data in lipidomicR
}
