% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_env.R
\name{lipidomicsR_env}
\alias{lipidomicsR_env}
\title{Global environment settings of LipidomicsR}
\usage{
lipidomicsR_env()
}
\value{
No return value, called for loading environment.
}
\description{
Loading environment for LipidomicsR. Please do not call it directly.
}
