% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaner.R
\name{nor.relative}
\alias{nor.relative}
\title{nor.relative()}
\usage{
nor.relative(
  data,
  Inlabel = c("PE(17:0/17:0)", "PC(17:0/17:0)", "PS(14:0/14:0)"),
  qc = 1:5,
  relative_as_default = TRUE,
  relative.mannual = NULL
)
}
\arguments{
\item{data}{Data frame, row lipidomics data.}

\item{Inlabel}{Inlabel Vector. Name of Internal label. Default as c('PE(17:0/17:0)','PC(17:0/17:0)','PS(14:0/14:0)')}

\item{qc}{Numeric value indicating the columns of quality control input.}

\item{relative_as_default}{Logical, default as TRUE for automatically searching for internal label data.}

\item{relative.mannual}{Vector, the exact channel name (if you want to define the channel of internal label mannually).}
}
\value{
Parameters for relative normalization
}
\description{
A function to calculate parameters for relative normalization.
}
