% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaner.R
\name{normalization_calculator}
\alias{normalization_calculator}
\title{normalization_calculator()}
\usage{
normalization_calculator(data, normalization.mode, normalization.index, qc=1:5)
}
\arguments{
\item{data}{Data frame. Row lipidomic data.}

\item{normalization.mode}{Vector. 'relative' for relative normalization. 'absolute' for absolute normalization.}

\item{normalization.index}{Data frame. normalization parameters, Suggested to be calculated by noridx.ouput()}

\item{qc}{Numeric vector. column of quality control. Default as 1:5.}
}
\value{
A data frame of normalized data if either 'relative' or 'absolute' mode is used. A list if both of them are used.
}
\description{
A function uses normalization parameters to calculate normalized lipidomic data.
}
