% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaner.R
\name{paramWizard}
\alias{paramWizard}
\title{paramWizard()}
\usage{
paramWizard()
}
\value{
A list of parameters.
}
\description{
A user-friendly interaction interface to input parameters for cleanXpert()
}
\details{
In this function, you should provide the following information:
\enumerate{
\item Path of row data.
\item Path of the paramters for your isotope label.
\item Whether default internal label should be used.
\item (optional) Name of internal label.
\item qc column and omitted qc column.
\item Whether replicate with the largest shifting should be removed.
\item (optional) The number of repeats of each group and repeats that should be removed within each group.
The parameters could either be import via key board typing or via external .txt file.
For first usage, key board typing input is suggested, and a .txt file will be automatically generated.
}
}
