% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaner.R
\name{percent.calculator}
\alias{percent.calculator}
\title{percent.calculator()}
\usage{
percent.calculator(data, absolute.dataset, internal.label, qc = 1:5)
}
\arguments{
\item{data}{Data frame. The row lipidomic data.}

\item{absolute.dataset}{Data frame. Normalization parameters for absolute
normalization. suggested to be generated by noridx().}

\item{internal.label}{Vector. Contains either the row channel name or lipid name of any external lipid labels.}

\item{qc}{Numeric. The column of quality control.}
}
\value{
Return a data frame of normalized lipidomic data in the percentage of lipid content.
}
\description{
A function to calculate the proportion of each lipid content.
}
