% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaner.R
\name{relative_calculator}
\alias{relative_calculator}
\title{relative.calculator()}
\usage{
relative_calculator(data, relative.dataset, qc = 1:5)
}
\arguments{
\item{data}{Row data of lipidomics}

\item{relative.dataset}{Data frame. Normalization index.}

\item{qc}{Numeric vector. The column of quality control.}
}
\value{
Return a data frame of relative normalized lipidomic data.
}
\description{
Internal function of lipidmicR::normalization_calculator()
}
