% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaner.R
\name{toGroup.calculator}
\alias{toGroup.calculator}
\title{toGroup.calculator()}
\usage{
toGroup.calculator(
  data,
  group,
  to,
  delete.pattern = c("_\\\\d", "(\\\\+)AcO", "_n", "\\\\(\\\\d+\\\\)")
)
}
\arguments{
\item{data}{Data frame. The row lipidomic data.}

\item{group}{Vector. The group information, recommended to be generated with groupXpert()}

\item{to}{Character. The group to be normalized to.}

\item{delete.pattern}{Pattern of characters that needs to be removed.}
}
\value{
A dataframe normalized to specified group.
}
\description{
toGroup.calculator()
}
