% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{total.abundance}
\alias{total.abundance}
\title{Calculate Total Abundance Data}
\usage{
total.abundance(data)
}
\arguments{
\item{data}{The data frame containing abundance data.}
}
\value{
A data frame containing total abundance of different lipid type, carbon number, and unsaturation.
}
\description{
This function calculates total abundance data based on specified groups.
}
