% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaner.R
\name{lEa}
\alias{lEa}
\title{lEa()}
\usage{
lEa(
  data,
  Inchannel = c("PC(15:0/18:1(d7))+AcO_1", "PE(15:0/18:1(d7))_2", "PS(15:0/18:1(d7))_2",
    "PG(15:0/18:1(d7))_2", "PI(15:0/18:1(d7))_2", "PA(15:0/18:1(d7))_2",
    "LPC18:1(d7)+AcO", "LPE18:1(d7)"),
  qc = 1:5,
  qc.omit = 1
)
}
\arguments{
\item{data}{Data frame. Row lipidomics data. Should be imported with importer().}

\item{Inchannel}{Vector. Exact channel name for radio label data.}

\item{qc}{Vector. Column of quantity control. Default as c(1:5).}

\item{qc.omit}{Vector. Column of quantity control that need to be omitted, default as 1.}
}
\value{
Data of internal label.
}
\description{
Internal function to extract label data in lipidomicR
}
