% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaner.R
\name{nor.absolute}
\alias{nor.absolute}
\title{nor.absolute()}
\usage{
nor.absolute(data, radio.data = NULL, sample = 2:5)
}
\arguments{
\item{data}{Data frame, row lipidomics data.}

\item{radio.data}{Data frame. Characteristic of the radio label data.
The row name must be the exact channel name of the label. Molecular mass should be
provided in a column named "Mass". Concentration should be provided in a column named "Concentration(mg/ml)".}

\item{sample}{Vector. Column of sample added with isotope label. Default as 2:5.}
}
\value{
Parameters for absolute normalization
}
\description{
A function to calculate parameters for absolute normalization.
}
