% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaner.R
\name{normalization_calculator}
\alias{normalization_calculator}
\title{normalization_calculator()}
\usage{
normalization_calculator(
  data,
  normalization.mode,
  normalization.index,
  delete.pattern = c("_\\\\d", "(\\\\+)AcO", "_n", "\\\\(\\\\d+\\\\)"),
  group,
  to
)
}
\arguments{
\item{data}{Data frame. Row lipidomic data.}

\item{normalization.mode}{Vector. Options can be 'absolute', 'relative', 'TSA', 'toGroup'.}

\item{normalization.index}{Data frame. normalization parameters, Suggested to be calculated by noridx.ouput()}

\item{delete.pattern}{Pattern of characters that needs to be removed.}

\item{group}{Vector. The group information, recommended to be generated with groupXpert()}

\item{to}{Character. The group to be normalized to.}
}
\value{
A data frame of normalized data if either 'relative' or 'absolute' mode is used. A list if both of them are used.
}
\description{
A function uses normalization parameters to calculate normalized lipidomic data.
}
